chore: organize test routes
