+++ packages/rsc/examples/basic/src/components/nav-menu.tsx
@@ -4,16 +4,13 @@ export function NavMenu() {
   return (
     <ul className="flex flex-col items-start gap-1 ml-5 list-disc">
       <li className="antd-link">
-        <Link href="/">/</Link>
+        <Link href="/test">/test</Link>
       </li>
       <li className="antd-link">
-        <Link href="/other">/other</Link>
+        <Link href="/test/other">/test/other</Link>
       </li>
       <li className="antd-link">
-        <Link href="/not-found">/not-found</Link>
-      </li>
-      <li className="antd-link">
-        <Link href="/demo/waku_02">/demo/waku_02</Link>
+        <Link href="/test/not-found">/test/not-found</Link>
       </li>
     </ul>
   );
