+++ packages/rsc/examples/basic/src/routes/layout.tsx
@@ -1,14 +1,21 @@
 import { Header } from "../components/header";
-import { NavMenu } from "../components/nav-menu";
-
-// TODO: move test pages under routes/test
+import { Link } from "../components/link";
 
 export default async function Layout(props: React.PropsWithChildren) {
   return (
     <div className="p-4 flex flex-col gap-2">
       <Header />
-      <NavMenu />
-      <input className="antd-input w-sm px-2" placeholder="test-input" />
+      <ul className="flex flex-col items-start gap-1 ml-5 list-disc">
+        <li className="antd-link">
+          <Link href="/">/</Link>
+        </li>
+        <li className="antd-link">
+          <Link href="/test">/test</Link>
+        </li>
+        <li className="antd-link">
+          <Link href="/demo/waku_02">/demo/waku_02</Link>
+        </li>
+      </ul>
       {props.children}
     </div>
   );
