+++ packages/rsc/examples/basic/src/routes/test/other/page.tsx
